moved scanner utils to their own file
