+++ src/server/frontend_wayland/generator/CMakeLists.txt
@@ -3,7 +3,8 @@ pkg_check_modules(XMLPP libxml++-2.6 REQUIRED)
 include_directories(SYSTEM ${XMLPP_INCLUDE_DIRS})
 
 add_executable(wrapper-generator
-  wrapper_generator.cpp     wrapper_generator.h
+  wrapper_generator.cpp
+  utils.cpp                 utils.h
   argument.cpp              argument.h
   method.cpp                method.h
   interface.cpp             interface.h
